re-add build:sitemap script
